requirements.txt
flask>=3.0
requests>=2.31
google-genai
python-dotenv
